03. Data Engineering/01. Environment & RDBMS/ds-sa-db-api-main/requirements.txt
attrs==21.2.0
certifi==2021.5.30
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pymongo==3.11.3
pyparsing==2.4.7
pytest==6.2.2
toml==0.10.2
urllib3==1.26.3
